Created test to confirm that error doesn't occur when executed from endpoint.

+++ swagger/restapi/operations/get_employee_salary_statement_urlbuilder.go
@@ -15,8 +15,8 @@ import (
 
 // GetEmployeeSalaryStatementURL generates an URL for the get employee salary statement operation
 type GetEmployeeSalaryStatementURL struct {
-	Month int32
-	Year  int32
+	Month int64
+	Year  int64
 
 	_basePath string
 	// avoid unkeyed usage
@@ -49,12 +49,12 @@ func (o *GetEmployeeSalaryStatementURL) Build() (*url.URL, error) {
 
 	qs := make(url.Values)
 
-	monthQ := swag.FormatInt32(o.Month)
+	monthQ := swag.FormatInt64(o.Month)
 	if monthQ != "" {
 		qs.Set("month", monthQ)
 	}
 
-	yearQ := swag.FormatInt32(o.Year)
+	yearQ := swag.FormatInt64(o.Year)
 	if yearQ != "" {
 		qs.Set("year", yearQ)
 	}
